rotation estimation/requirements.txt
pyrender==0.1.45
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.5
scikit-image==0.18.1
trimesh==3.9.9
scipy==1.5.1
Pillow==7.2.0
imageio==2.9.0
numpy==1.19.5
structlog==21.1.0
matplotlib==3.3.4
tqdm==4.59.0
imgaug==0.4.0
opencv-python==4.5.1.48
torch==1.8.0+cu111
torchvision==0.9.0+cu111
PyYAML==5.4.1
tensorboard==2.4.1
cudatoolkit==11.1
Blender
detectron2
pytorch3d
